% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{assign_covariables}
\alias{assign_covariables}
\title{Assign dependent covariables to a design table}
\usage{
assign_covariables(design_lab, design_keys, covariables, lang = NULL)
}
\description{
Assign dependent covariables to a design table
}
\keyword{internal}
