% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_checker.R
\name{build_design_table_html}
\alias{build_design_table_html}
\title{Build HTML table from preview data}
\usage{
build_design_table_html(preview_df)
}
\description{
Build HTML table from preview data
}
\keyword{internal}
