% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_checker.R
\name{build_preview_vignette_data}
\alias{build_preview_vignette_data}
\title{Build preview vignette data structure from design_data}
\usage{
build_preview_vignette_data(vig_data, langs)
}
\description{
Build preview vignette data structure from design_data
}
\keyword{internal}
