% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{detect_languages}
\alias{detect_languages}
\title{Detect available languages from factors structure}
\usage{
detect_languages(factors)
}
\arguments{
\item{factors}{Named list of factor definitions}
}
\value{
Character vector of detected language codes
}
\description{
Detect available languages from factors structure
}
