% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{get_factor_labels}
\alias{get_factor_labels}
\title{Helper: Get factor labels for one language}
\usage{
get_factor_labels(factor_def, lang = NULL)
}
\description{
Helper: Get factor labels for one language
}
\keyword{internal}
