% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{get_lang}
\alias{get_lang}
\title{Helper: Get language-specific text from multilingual object}
\usage{
get_lang(x, lang, fallback = NA_character_)
}
\description{
Helper: Get language-specific text from multilingual object
}
\keyword{internal}
