% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{label_design_factors}
\alias{label_design_factors}
\title{Store design dimensions as labelled factors}
\usage{
label_design_factors(design_data, design_keys, factors, lang = NULL)
}
\description{
Store design dimensions as labelled factors
}
\keyword{internal}
