% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{normalize_covariables}
\alias{normalize_covariables}
\title{Normalize dependent covariable definitions}
\usage{
normalize_covariables(covariables = NULL)
}
\description{
Normalize dependent covariable definitions
}
\keyword{internal}
