% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{normalize_design}
\alias{normalize_design}
\title{Normalize design input from data frame or list}
\usage{
normalize_design(design = NULL)
}
\description{
Normalize design input from data frame or list
}
\keyword{internal}
