% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{prefix_vignette_placeholders}
\alias{prefix_vignette_placeholders}
\title{Prefix vignette placeholders in text}
\usage{
prefix_vignette_placeholders(texts, keywords, prefix = "vig", num)
}
\arguments{
\item{texts}{Character vector of texts with placeholders in \verb{\\\\\{...\\\\\}}.}

\item{keywords}{Character vector of variable names to prefix}

\item{prefix}{Prefix string (default: "vig")}

\item{num}{Vignette number to use in prefix}
}
\value{
Character vector with prefixed variable names
}
\description{
Prefix vignette placeholders in text
}
\examples{
txt <- "{name} is {age}"
prefix_vignette_placeholders(txt, c("name", "age"), prefix = "vig", num = 1)

}
\keyword{internal}
