% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{prepare_vignettes_single_lang}
\alias{prepare_vignettes_single_lang}
\title{Generate vignette codes for a single language}
\usage{
prepare_vignettes_single_lang(
  factors,
  design_df = NULL,
  text = NULL,
  vig_n = NULL,
  lang = NULL,
  prefix = "vig",
  covariables = NULL,
  store_design_data = TRUE
)
}
\description{
Generate vignette codes for a single language
}
\keyword{internal}
