% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_builder.R
\name{vignettes_to_seed}
\alias{vignettes_to_seed}
\title{Add vignettes to a seed structure}
\usage{
vignettes_to_seed(vig_data, seed, deckgroupsuffix = "intro")
}
\arguments{
\item{vig_data}{Output from build_vignette_data (with YAML format)}

\item{seed}{LimeSeed seed object}

\item{deckgroupsuffix}{Suffix of the group containing the deck randomizer.
Set to \code{NULL} if the deck variable is managed elsewhere.}
}
\value{
Modified seed object with vignette defaults and hidden variables
added to the corresponding groups.
}
\description{
Add vignettes to a seed structure
}
